Add handler and UppdateDataBase

+++ Services/RestaurantServies.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
+using RestaurantApi.Authorization;
 using RestaurantApi.Models;
 using RestaurantAPI.Exceptions;
 using RestaurantApii.DataBase;
@@ -8,6 +10,7 @@ using RestaurantApii.Entities;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace RestaurantApi.Services
@@ -16,9 +19,9 @@ namespace RestaurantApi.Services
     {
         RestaurantDto GetById(int id);
         IEnumerable<RestaurantDto> GetAll();
-        int Create(CreateRestaurantDto dto);
-        void Delete(int id);
-        void Modify(int id, ModifyRestaurantDto dto);
+        int Create(CreateRestaurantDto dto, int userId);
+        void Delete(int id, ClaimsPrincipal user);
+        void Modify(int id, ModifyRestaurantDto dto, ClaimsPrincipal user);
 
     }
     public class RestaurantServies : IRestaurantServies
@@ -26,11 +29,13 @@ namespace RestaurantApi.Services
         private readonly RestaurantDbContext _dbContext;
         private readonly IMapper _mapper;
         private readonly ILogger<RestaurantServies> _logger;
-        public RestaurantServies(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantServies> logger)
+        private readonly IAuthorizationService _authorizationService;
+        public RestaurantServies(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantServies> logger, IAuthorizationService authorizationService)
         {
             _dbContext = dbContext;
             _mapper = mapper;
             _logger = logger;
+            _authorizationService = authorizationService;
         }
         //Pobiebanie Restauracji Po wskazanym ID
         public RestaurantDto GetById(int id )
@@ -62,9 +67,10 @@ namespace RestaurantApi.Services
         }
 
        //Twozenie nowej Restauracji 
-        public int Create(CreateRestaurantDto dto)
+        public int Create(CreateRestaurantDto dto, int userId)
         {
             var restaurant = _mapper.Map<Restaurant>(dto);
+            restaurant.CreatedById = userId;
             _dbContext.Restaurants.Add(restaurant);
             _dbContext.SaveChanges();
 
@@ -72,7 +78,7 @@ namespace RestaurantApi.Services
         }
 
         //Usuwanie Restauracji z Adressem 
-        public void Delete(int id)
+        public void Delete(int id, ClaimsPrincipal user)
         {
             _logger.LogTrace($"Restaurant with id: {id} DELETE action invoked");
 
@@ -86,20 +92,33 @@ namespace RestaurantApi.Services
             {
                 throw new NotFoundException("Restaurant not found");
             }
+            var wynik = _authorizationService.AuthorizeAsync(user, restaurant, new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+            if (!wynik.Succeeded)
+            {
+                throw new ForbidException();
+            }
             _dbContext.Restaurants.Remove(restaurant);
             _dbContext.Adresses.Remove(address);
             _dbContext.SaveChanges();
         }
     
-        public void Modify(int id, ModifyRestaurantDto dto)
+        public void Modify(int id, ModifyRestaurantDto dto, ClaimsPrincipal user)
         {
             var restaurant = _dbContext
                 .Restaurants
                 .FirstOrDefault(x => x.Id == id);
+
             if (restaurant is null)
             {
                 throw new NotFoundException("Restaurant not found");
             }
+
+            var wynik = _authorizationService.AuthorizeAsync(user, restaurant, new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+            if (!wynik.Succeeded)
+            {
+                throw new ForbidException();
+            }
+
             restaurant.Name = dto.Name;
             restaurant.Description = dto.Description;
             restaurant.HasDelivery = dto.HasDelivery;
